+++ src/components/board.js
@@ -1,0 +1,6 @@
+const createBoardTemplate = () =>
+  `<section class="board container">
+    <div class="board__tasks"></div>
+  </section>`;
+
+export {createBoardTemplate};
